pulp_soc/soc_clk_rst_gen: Do four phase handshake for clk div access
Instead of doing the currently ad-hoc approach of keeping pready up
until the the master de-asserts penable, we consider the soc to be
living in a synchronous clock domain and do a proper four phase
handshake do transmit the apb request to the clock divider (which lives
in the sys_clk domain).

To do that, a state machine clocked by the soc clock accepts the apb
request and translates it into a four phase handshake with a state
machine living in the sys_clk domain which in turn does a ready/valid
transaction with the clock divider interface.